+++ java/CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.9.3
-*
+* Add Apache Sling DavEx Servlet OSGI config support #11
 
 ### 0.9.2
 * Add operation parameter support to sling#postAgentWithHttpInfo